Update: Fix computed and test bugs
