+++ packages/reports/addon/components/filter-builders/date-dimension.js
@@ -44,7 +44,7 @@ export default class DateDimensionFilterBuilderComponent extends BaseFilterBuild
    * @property {Object} filter
    * @override
    */
-  @computed('requestFragment.{operator,dimension,values.[]}')
+  @computed('supportedOperators', 'requestFragment.{operator,dimension,values.[]}')
   get filter() {
     const { requestFragment } = this;
     const serializedFilter =
