+++ packages/reports/ember-cli-build.js
@@ -16,8 +16,5 @@ module.exports = function(defaults) {
    * behave. You most likely want to be modifying `./index.js` or app's build file
    */
 
-  //needed for our helper `hbs-with-modal`
-  app.import('vendor/ember/ember-template-compiler.js');
-
   return app.toTree();
 };
