+++ packages/reports/package.json
@@ -16,13 +16,14 @@
   },
   "scripts": {
     "build": "ember build --environment=production",
-    "lint:js": "eslint . && tsc --allowJs false --noEmit",
-    "lint": "npm-run-all --aggregate-output --continue-on-error --parallel lint:*",
     "lint:hbs": "ember-template-lint .",
+    "lint:js": "eslint .",
+    "lint:ts": "tsc --allowJs false --noEmit",
+    "lint": "npm-run-all --aggregate-output --continue-on-error --parallel lint:*",
     "start": "ember serve",
     "test": "npm-run-all lint:* test:*",
     "test:ember": "ember test",
-    "test:ember-compatibility": "ember try:each",
+    "try:ember-compatibility": "ember try:each",
     "prepublishOnly": "ember ts:precompile",
     "postpublish": "ember ts:clean"
   },
@@ -108,7 +109,7 @@
     "ember-resize": "^0.3.4",
     "ember-resolver": "^8.0.0",
     "ember-sortable": "2.1.3",
-    "ember-source": "~3.19.0",
+    "ember-source": "~3.12.0",
     "ember-source-channel-url": "^2.0.1",
     "ember-template-compiler": "^1.8.0",
     "ember-template-lint": "^2.8.0",
