+++ packages/reports/tests/helpers/get-date-range.js
@@ -1,0 +1,9 @@
+import { formatDateRange } from 'navi-reports/helpers/format-interval-inclusive-inclusive';
+import { getIsoDateTimePeriod } from 'navi-core/utils/date';
+
+export function getDateRangeFormat(source) {
+  const dateTimePeriod = source.request.logicalTable.timeGrain;
+  const { start, end } = source.filter.values.firstObject.asMomentsForTimePeriod(dateTimePeriod);
+  end.subtract(1, getIsoDateTimePeriod(dateTimePeriod));
+  return formatDateRange(start, end, dateTimePeriod);
+}
