+++ packages/reports/tests/integration/components/filter-values/lookback-input-test.js
@@ -4,7 +4,7 @@ import { render, fillIn, triggerEvent, click, findAll } from '@ember/test-helper
 import { A as arr } from '@ember/array';
 import hbs from 'htmlbars-inline-precompile';
 import Interval from 'navi-core/utils/classes/interval';
-import { getDateRangeFormat } from './date-range-test';
+import { getDateRangeFormat } from '../../../helpers/get-date-range';
 import { clickTrigger } from 'ember-basic-dropdown/test-support/helpers';
 import $ from 'jquery';
